WIP - trying to see if using post-command instead of activate-mark/deactivate-mark hooks is better
There are times when I want a command to end in visual state, but it doesn't.  Specifically commands that modify the buffer.  Explicitly activating the mark at the end of the command doesn't seem to help.  This change doesn't fix it.  But I imagine this could be better for various reasons.  But I'm uncertain.